+++ src/main/java/com/cedarsoftware/util/FastReader.java
@@ -3,6 +3,27 @@ package com.cedarsoftware.util;
 import java.io.IOException;
 import java.io.Reader;
 
+/**
+ * Buffered, Pushback, Reader that does not use synchronization.   Much faster than the JDK variants because
+ * they use synchronization.  Typically, this class is used with a separate instance per thread, so
+ * synchronization is not needed.
+ * <p>
+ * @author John DeRegnaucourt (jdereg@gmail.com)
+ *         <br>
+ *         Copyright (c) Cedar Software LLC
+ *         <br><br>
+ *         Licensed under the Apache License, Version 2.0 (the "License");
+ *         you may not use this file except in compliance with the License.
+ *         You may obtain a copy of the License at
+ *         <br><br>
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *         <br><br>
+ *         Unless required by applicable law or agreed to in writing, software
+ *         distributed under the License is distributed on an "AS IS" BASIS,
+ *         WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *         See the License for the specific language governing permissions and
+ *         limitations under the License.
+ */
 public class FastReader extends Reader {
     private Reader in;
     private char[] buf;