+++ src/test/java/com/cedarsoftware/util/TestUtilTest.java
@@ -2,11 +2,6 @@ package com.cedarsoftware.util;
 
 import org.junit.jupiter.api.Test;
 
-import java.net.URL;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-
 /**
  * @author John DeRegnaucourt (jdereg@gmail.com)
  *         <br>
@@ -55,19 +50,5 @@ public class TestUtilTest
         assert TestUtil.checkContainsIgnoreCase("This is the source string to test.", "Source", "string", "Test");
         assert !TestUtil.checkContainsIgnoreCase("This is the source string to test.", "Source", "string", "Text");
         assert !TestUtil.checkContainsIgnoreCase("This is the source string to test.", "Test", "Source", "string");
-    }
-
-    public static String fetchResource(String name)
-    {
-        try
-        {
-            URL url = TestUtil.class.getResource("/" + name);
-            Path resPath = Paths.get(url.toURI());
-            return new String(Files.readAllBytes(resPath));
-        }
-        catch (Exception e)
-        {
-            throw new RuntimeException(e);
-        }
     }
 }